requirements.txt
appdirs==1.4.4
black==20.8b1
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
gspread==3.6.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
oauth2client==4.1.3
oauthlib==3.1.0
pathspec==0.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.15.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1